refactored to contain all react30 challenges separated into individual pages/components
